\name{zipcode-package}
\alias{zipcode-package}
\docType{package}
\title{U.S. ZIP code database for geocoding}
\description{This package contains a database of city, state,
  latitude, and longitude information for U.S. ZIP codes from the CivicSpace
  Database (August 2004) and augmented by Daniel Coven's 
  federalgovernmentzipcodes.us web site (updated January 22, 2012).
  Previous versions of this package (before 1.0) were based solely 
  on the CivicSpace data, so an original version of the CivicSpace 
  database is also included.
  }
\details{
\tabular{ll}{Package: \tab zipcode\cr
Maintainer: \tab Jeffrey Breen <jeffrey@atmosgrp.com>\cr
Author: \tab Jeffrey Breen <jeffrey@atmosgrp.com>\cr
Version: \tab 1.0\cr
License: \tab CC BY-SA 2.0 + file LICENSE\cr
Title: \tab U.S. ZIP code database for geocoding\cr}
}
\author{Jeffrey Breen <jeffrey@atmosgrp.com>}

\keyword{ package }


