\name{zipcode-package}
\alias{zipcode-package}
\docType{package}
\title{U.S. ZIP Code database}
\description{This package contains the "CivicSpace US ZIP Code Database" by Schuyler Erle,
August 2004 as a data.frame.  Believed to contain over 98\% of U.S. ZIP Codes as of 2004
The source data is available in CSV from \url{http://www.boutell.com/zipcodes/}
and \url{http://mappinghacks.com/data/} and has been released under the Creative
Commons Share-Alike license. What little code is included in this package to clean
up suspect ZIP codes is released under GPL-2.}
\details{
\tabular{ll}{Package: \tab zipcode\cr
Maintainer: \tab Jeffrey Breen <jbreen@cloudsmiths.com>\cr
Author: \tab Jeffrey Breen <jbreen@cloudsmiths.com>\cr
Version: \tab 0.1\cr
License: \tab GPL-2 + Creative Commons Share-Alike\cr
Title: \tab U.S. ZIP Code database\cr}
}
\author{Jeffrey Breen <jbreen@cloudsmiths.com>}

\keyword{ package }


